Update to latest Slicer preview release
Fix compatibility issue with latest mybinder: mybinder now uses a complex logic in the docker command to determine jupyter classic or lab interface should be used and it is hardcoded to use system Python.
We now intercept the system Python call, capture the jupyter notebook arguments, and run Slicer's notebook with it.